TTALGO-1127 Backtester, moving average bot: Close volume = 0 TTALGO-1104 Backtester, moving average bot: Volumes aren't calculated

+++ TickTrader.Algo.Api/Ext/DealerEmulator.cs
@@ -20,7 +20,7 @@ namespace TickTrader.Algo.Api.Ext
         RateUpdate CurrentRate { get; }
 
         void Confirm();
-        void Confirm(double amount, double price);
+        void Confirm(decimal amount, double price);
         void Reject();
     }
