+++ TickTrader.Algo.Core/Entities/NetPositionOpenInfo.cs
@@ -16,8 +16,8 @@ namespace TickTrader.Algo.Core
 
     internal class NetPositionCloseInfo
     {
-        public double CloseAmount { get; set; }
-        public double ClosePrice { get; set; }
-        public double BalanceMovement { get; set; }
+        public decimal CloseAmount { get; set; }
+        public decimal ClosePrice { get; set; }
+        public decimal BalanceMovement { get; set; }
     }
 }
